Animations, Middle and Top level of castle outlined.
All of the animations for the player and enemies are in, and any extra animations are set up and ready to be made whenever we will need them, such as blood or the mansion final boss.  I outlined the middle and top floors of the castle level, and started the top castle level detailing.